requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
pycryptodome==3.11.0
qrcode==7.3.1
Werkzeug==2.0.2
